Fixes for incorrect paths due to using back slashes instead of forward slashes.
This is the 5th patch 'flightgear-fgaddon_path_fix5_backslash.patch' attached to
the thread at https://sourceforge.net/p/flightgear/mailman/message/33744249/ .

This change is a bit different from the other patches of the series as it is not
due to the FGData-FGAddon split, but was found when testing aircraft for
incorrect paths.  This replaces the back slashes which prevent certain XML files
from being loaded with functional forward slashes for the aircraft 767-300,
A340-600, Cessna337, CRJ-200, and Dromader.



git-svn-id: http://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/A340-600@545 9eb9bc84-db9d-4e05-893b-26dd41ac94a6